feat(javascript-node-14): update definition for Node 14
This changes details of the base `javascript-node-*` template to add
support for Node 14. That way, if accepted, other depending templates
could be updated, if relying on that base image

No specific changes compared to v12 were made. The 'forwardPorts' option
is left-as-is, as it's documented in the template README

Thanks!