Update Puppet to point to official docker image
This commit updates the Puppet Dockerfile to point to the official pdk
docker image at `puppet/pdk` by replacing the custom Dockerfile with the
pdk image. This is essentially the same end result, as this custom
Dockerfile was based on the official repo. It leaves a commented block as
an example of how to further customize the image.