chore: changing "restart" to "start" of sshd server
pid file of the sshd server gets saved in the container. When it is restarted on a different VM, the pid is different and that process gets killed.
This results in random process kills and failure of about 10% of the resumes.
Fix is to change "restart" to "start" so that no process gets killed as container doesn't have any processes at that time.