Include Python virtual env name on the terminal prompt
Normally when you activate a Python virtual env, it updates the terminal prompt to include the env name wrapped in parenthesis as a prefix on the PS1. Currently, that doesn't work with the default bash `codespaces` theme. This change updates the theme to prepend the virtual env name to the prompt if a virtual env is defined via the `$VIRTUAL_ENV` environment variable, which is similar to the logic that the venv activate script runs.