Added other MySQL Environment Variables
We set the hostname out of necessity as commonly development configurations reference MySQL running on localhost. It might be nice to include other variables for an application server to pickup. So I added commonly used MYSQL_{USER,PASSWORD,DATABASE}. Should we also include one for the port or assume most applications in development mode comply with TCP:3306?