Add inotify-tools to elixir phoenix postgres template
inotify-tools is required in linux environments for live code reloading with Phoenix. Users seeking to use this template to development will be met with a warning when starting a phoenix project for the first time and will need to install it manually. Adding it as a dependency to the dev container would avoid that and help people get going quicker.

ref: https://hexdocs.pm/phoenix/installation.html#inotify-tools-for-linux-users