refactor: download bazelisk to /usr/local/bin/bazel
The original implementation of this links `bazelisk` to `/user/local/bin/bazelisk`, but that means a user would have to type `bazelisk` to have it install bazel. The more natural way for a user to interact in a `bazel` project is just to run `bazel`. If `bazelisk` is linked to the `bazel` path it will automatically fetch `bazel` for the end user and not have to worry about different commands.